move app.jsx and app.css to /App
